Add 5GLA API integration functionality
Added functionality to interact with 5GLA API in the application by creating a new ApiIntegrationService class which checks the availability of the API and integrated this with the ConfigManager class. Also added proper tests for 5GLA API integration, and added requests to the list of requirements in order to perform HTTP requests to the API. Also, key error handling is now supported when there's an attempt to access an unavailable key in the configuration file or environment variables.