Add api_image_endpoint to configuration
In this commit, the new endpoint 'api_image_endpoint' is added to the configuration in app/config/config/config.yaml. This endpoint (/mica-sense/images/) is necessary for services that need to access or manipulate images through the API. The corresponding integration test file 'test_5gla_api_integration_service.py' is also updated in this commit.