Refactor directory structure from 'app' to 'src'
Transferred all python scripts and resources from 'app' directory to 'src'. Updating of respective import statements and relative file paths have been done in several files, including integration and config tests, as a part of refactoring. This change was implemented to bring consistency and clarity to the overall codebase structure.