Add config setting for image sending and enhance some integrations
The image sending functionality can now be toggled on or off by a configuration setting. The intervals between sending have been generalized, not only for images, and moved to the configuration. Additionally, latitude and longitude are now converted to degrees in the camera service. The HTTP requests in the API integration service now include a 'Content-Type' in their headers and