Remove API availability check
The API availability check functionality has been removed from the code. The application was previously checking if the API is available before starting the process. This has been deemed unnecessary and reduced the efficiency of the program. Now, the program directly starts the process, saving time and resources. Additionally, some code related to the functionality has been removed and config file adjusted accordingly.